Rename time range selection group of classes
