+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/ranges/data/TodayRangeData.kt
@@ -5,7 +5,7 @@ import com.woocommerce.android.extensions.formatToMMMddYYYY
 import com.woocommerce.android.extensions.oneDayAgo
 import com.woocommerce.android.extensions.startOfCurrentDay
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRange
-import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRangeData
+import com.woocommerce.android.ui.analytics.ranges.StatsTimeRangeData
 import java.util.Calendar
 import java.util.Date
 import java.util.Locale
@@ -23,7 +23,7 @@ class TodayRangeData(
     referenceDate: Date,
     locale: Locale,
     referenceCalendar: Calendar
-) : AnalyticsHubTimeRangeData(referenceCalendar) {
+) : StatsTimeRangeData(referenceCalendar) {
     override val currentRange: AnalyticsHubTimeRange
     override val previousRange: AnalyticsHubTimeRange
     override val formattedCurrentRange: String
