+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/main/MainNavigationRouter.kt
@@ -1,7 +1,7 @@
 package com.woocommerce.android.ui.main
 
 import android.view.View
-import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubDateRangeSelection
+import com.woocommerce.android.ui.analytics.ranges.StatsTimeRangeSelection
 
 interface MainNavigationRouter {
     fun isAtNavigationRoot(): Boolean
@@ -51,5 +51,5 @@ interface MainNavigationRouter {
     fun showFeedbackSurvey()
     fun showSettingsScreen()
 
-    fun showAnalytics(targetPeriod: AnalyticsHubDateRangeSelection.SelectionType)
+    fun showAnalytics(targetPeriod: StatsTimeRangeSelection.SelectionType)
 }
